Recovery duration for DC is taken from the config
Now, the recovery duration is specified in microseconds.

+++ modules/bcb/zephyr/include/lib/bcb_tc_def_msm.h
@@ -63,7 +63,7 @@ typedef struct bcb_msm_config {
 	bcb_tc_def_msm_csom_t csom; /**< Closed-state operation mode. */
 	bool rec_enabled; /**< Set to true if recovery is enabled. */
 	uint16_t rec_attempts; /**< Number of recovery attempts. */
-	uint16_t rec_delay; /**< Recovery delay in milliseconds. */
+	uint32_t rec_delay; /**< Recovery delay in microseconds. */
 } bcb_tc_def_msm_config_t;
 
 /**
